// frontend/src/lib/protocol/generated/PublishedExternalDevice.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ExternalDeviceId } from "./ExternalDeviceId";

export type PublishedExternalDevice = { deviceId: ExternalDeviceId, enabled: boolean, } & ({ "type": "tcp", host: string, port: number, } | { "type": "bluetooth", address: string, });